@@ -1,43 +1,55 @@
-import React, {Component} from 'react';
-import {
-  FlatList,
-  ActivityIndicator,
-  View
-} from 'react-native';
-import Reservation from './reservation';
+import _ from 'lodash';
 import PropTypes from 'prop-types';
 import XDate from 'xdate';
-
+import React, {Component} from 'react';
+import {FlatList, ActivityIndicator, View} from 'react-native';
+import {extractComponentProps} from '../../component-updater';
 import dateutils from '../../dateutils';
 import styleConstructor from './style';
+import Reservation from './reservation';
 
-class ReactComp extends Component {
+class ReservationList extends Component {
+  static displayName = 'IGNORE';
+
   static propTypes = {
-    // specify your item comparison function for increased performance
-    rowHasChanged: PropTypes.func,
-    // specify how each item should be rendered in agenda
-    renderItem: PropTypes.func,
-    // specify how each date should be rendered. day can be undefined if the item is not first in that day.
-    renderDay: PropTypes.func,
-    // specify how empty date content with no items should be rendered
-    renderEmptyDate: PropTypes.func,
-    // callback that gets called when day changes while scrolling agenda list
-    onDayChange: PropTypes.func,
-    // onScroll ListView event
-    onScroll: PropTypes.func,
-    // the list of items that have to be displayed in agenda. If you want to render item as empty date
-    // the value of date key kas to be an empty array []. If there exists no value for date key it is
-    // considered that the date in question is not yet loaded
+    ...Reservation.propTypes,
+    /** the list of items that have to be displayed in agenda. If you want to render item as empty date
+    the value of date key kas to be an empty array []. If there exists no value for date key it is
+    considered that the date in question is not yet loaded */ 
     reservations: PropTypes.object,
-
     // specify a flag to show the left day on the left of reservation
     showDate: PropTypes.bool,
 
     selectedDay: PropTypes.instanceOf(XDate),
     topDay: PropTypes.instanceOf(XDate),
+    /** Show items only for the selected day. Default = false. */
+    showOnlySelectedDayItems: PropTypes.bool,
+    /** callback that gets called when day changes while scrolling agenda list. */ 
+    onDayChange: PropTypes.func,
+    /** specify what should be rendered instead of ActivityIndicator. */
+    renderEmptyData: PropTypes.func,
+
+    /** onScroll ListView event */
+    onScroll: PropTypes.func,
+    /** Called when the user begins dragging the agenda list. **/
+    onScrollBeginDrag: PropTypes.func,
+    /** Called when the user stops dragging the agenda list. **/
+    onScrollEndDrag: PropTypes.func,
+    /** Called when the momentum scroll starts for the agenda list. **/
+    onMomentumScrollBegin: PropTypes.func,
+    /** Called when the momentum scroll stops for the agenda list. **/
+    onMomentumScrollEnd: PropTypes.func,
+    /** A RefreshControl component, used to provide pull-to-refresh functionality for the ScrollView. */
     refreshControl: PropTypes.element,
+    /** Set this true while waiting for new data from a refresh. */
     refreshing: PropTypes.bool,
-    onRefresh: PropTypes.func,
+    /** If provided, a standard RefreshControl will be added for "Pull to Refresh" functionality. Make sure to also set the refreshing prop correctly. */
+    onRefresh: PropTypes.func
+  };
+
+  static defaultProps = {
+    refreshing: false,
+    selectedDay: XDate(true)
   };
 
   static defaultProps = {
@@ -46,17 +58,35 @@ class ReactComp extends Component {
 
   constructor(props) {
     super(props);
-    this.styles = styleConstructor(props.theme);
+
+    this.style = styleConstructor(props.theme);
+
     this.state = {
       reservations: []
     };
-    this.heights=[];
-    this.selectedDay = this.props.selectedDay;
+
+    this.heights = [];
+    this.selectedDay = props.selectedDay;
     this.scrollOver = true;
   }
 
-  componentWillMount() {
+  componentDidMount() {
     this.updateDataSource(this.getReservations(this.props).reservations);
+  }
+
+  componentDidUpdate(prevProps) {
+    if (prevProps !== this.props) {
+      if (!dateutils.sameDate(prevProps.topDay, this.props.topDay)) {
+        this.setState(
+          {
+            reservations: []
+          },
+          () => this.updateReservations(this.props)
+        );
+      } else {
+        this.updateReservations(this.props);
+      }
+    }
   }
 
   updateDataSource(reservations) {
@@ -66,8 +96,9 @@ class ReactComp extends Component {
   }
 
   updateReservations(props) {
+    const {selectedDay} = props;
     const reservations = this.getReservations(props);
-    if (this.list && !dateutils.sameDate(props.selectedDay, this.selectedDay)) {
+    if (this.list && !dateutils.sameDate(selectedDay, this.selectedDay)) {
       let scrollPosition = 0;
       for (let i = 0; i < reservations.scrollPosition; i++) {
         scrollPosition += this.heights[i] || 0;
@@ -75,63 +106,10 @@ class ReactComp extends Component {
       this.scrollOver = false;
       this.list.scrollToOffset({offset: scrollPosition, animated: true});
     }
-    this.selectedDay = props.selectedDay;
+    this.selectedDay = selectedDay;
     this.updateDataSource(reservations.reservations);
   }
-
-  componentWillReceiveProps(props) {
-    if (!dateutils.sameDate(props.topDay, this.props.topDay)) {
-      this.setState({
-        reservations: []
-      }, () => {
-        this.updateReservations(props);
-      });
-    } else {
-      this.updateReservations(props);
-    }
-  }
-
-  onScroll(event) {
-    const yOffset = event.nativeEvent.contentOffset.y;
-    this.props.onScroll(yOffset);
-    let topRowOffset = 0;
-    let topRow;
-    for (topRow = 0; topRow < this.heights.length; topRow++) {
-      if (topRowOffset + this.heights[topRow] / 2 >= yOffset) {
-        break;
-      }
-      topRowOffset += this.heights[topRow];
-    }
-    const row = this.state.reservations[topRow];
-    if (!row) return;
-    const day = row.day;
-    const sameDate = dateutils.sameDate(day, this.selectedDay);
-    if (!sameDate && this.scrollOver) {
-      this.selectedDay = day.clone();
-      this.props.onDayChange(day.clone());
-    }
-  }
-
-  onRowLayoutChange(ind, event) {
-    this.heights[ind] = event.nativeEvent.layout.height;
-  }
-
-  renderRow({item, index}) {
-    return (
-      <View onLayout={this.onRowLayoutChange.bind(this, index)}>
-        <Reservation
-          item={item}
-          renderItem={this.props.renderItem}
-          renderDay={this.props.renderDay}
-          renderEmptyDate={this.props.renderEmptyDate}
-          theme={this.props.theme}
-          rowHasChanged={this.props.rowHasChanged}
-          showDate={this.props.showDate}
-        />
-      </View>
-    );
-  }
-
+  
   getReservationsForDay(iterator, props) {
     const day = iterator.clone();
     const res = props.reservations[day.toString('yyyy-MM-dd')];
@@ -144,27 +122,28 @@ class ReactComp extends Component {
         };
       });
     } else if (res) {
-      return [{
-        date: iterator.clone(),
-        day
-      }];
+      return [
+        {
+          date: iterator.clone(),
+          day
+        }
+      ];
     } else {
       return false;
     }
   }
 
-  onListTouch() {
-    this.scrollOver = true;
-  }
-
   getReservations(props) {
-    if (!props.reservations || !props.selectedDay) {
+    const {selectedDay, showOnlySelectedDayItems} = props;
+    if (!props.reservations || !selectedDay) {
       return {reservations: [], scrollPosition: 0};
     }
+
     let reservations = [];
     if (this.state.reservations && this.state.reservations.length) {
       const iterator = this.state.reservations[0].day.clone();
-      while (iterator.getTime() < props.selectedDay.getTime()) {
+
+      while (iterator.getTime() < selectedDay.getTime()) {
         const res = this.getReservationsForDay(iterator, props);
         if (!res) {
           reservations = [];
@@ -175,44 +154,109 @@ class ReactComp extends Component {
         iterator.addDays(1);
       }
     }
+
     const scrollPosition = reservations.length;
-    const iterator = props.selectedDay.clone();
-    for (let i = 0; i < 31; i++) {
+    const iterator = selectedDay.clone();
+    if (showOnlySelectedDayItems) {
       const res = this.getReservationsForDay(iterator, props);
+
       if (res) {
-        reservations = reservations.concat(res);
+        reservations = res;
       }
       iterator.addDays(1);
+    } else {
+      for (let i = 0; i < 31; i++) {
+        const res = this.getReservationsForDay(iterator, props);
+
+        if (res) {
+          reservations = reservations.concat(res);
+        }
+        iterator.addDays(1);
+      }
     }
 
     return {reservations, scrollPosition};
   }
 
-  render() {
-    if (!this.props.reservations || !this.props.reservations[this.props.selectedDay.toString('yyyy-MM-dd')]) {
-      if (this.props.renderEmptyData) {
-        return this.props.renderEmptyData();
+  onScroll = event => {
+    const yOffset = event.nativeEvent.contentOffset.y;
+    _.invoke(this.props, 'onScroll', yOffset);
+
+    let topRowOffset = 0;
+    let topRow;
+    for (topRow = 0; topRow < this.heights.length; topRow++) {
+      if (topRowOffset + this.heights[topRow] / 2 >= yOffset) {
+        break;
       }
-      return (<ActivityIndicator style={{marginTop: 80}}/>);
+      topRowOffset += this.heights[topRow];
     }
+
+    const row = this.state.reservations[topRow];
+    if (!row) return;
+
+    const day = row.day;
+    const sameDate = dateutils.sameDate(day, this.selectedDay);
+    if (!sameDate && this.scrollOver) {
+      this.selectedDay = day.clone();
+      _.invoke(this.props, 'onDayChange', day.clone());
+    }
+  };
+
+  onListTouch() {
+    this.scrollOver = true;
+  }
+
+  onRowLayoutChange(ind, event) {
+    this.heights[ind] = event.nativeEvent.layout.height;
+  }
+
+  onMoveShouldSetResponderCapture = () => {
+    this.onListTouch();
+    return false;
+  };
+
+  renderRow = ({item, index}) => {
+    const reservationProps = extractComponentProps(Reservation, this.props);
+
+    return (
+      <View onLayout={this.onRowLayoutChange.bind(this, index)}>
+        <Reservation {...reservationProps} item={item} />
+      </View>
+    );
+  };
+
+  render() {
+    const {reservations, selectedDay, theme, style} = this.props;
+    if (!reservations || !reservations[selectedDay.toString('yyyy-MM-dd')]) {
+      if (_.isFunction(this.props.renderEmptyData)) {
+        return _.invoke(this.props, 'renderEmptyData');
+      }
+
+      return <ActivityIndicator style={this.style.indicator} color={theme && theme.indicatorColor} />;
+    }
+
     return (
       <FlatList
-        ref={(c) => this.list = c}
-        style={this.props.style}
-        contentContainerStyle={this.styles.content}
-        renderItem={this.renderRow.bind(this)}
+        ref={c => (this.list = c)}
+        style={style}
+        contentContainerStyle={this.style.content}
         data={this.state.reservations}
-        onScroll={this.onScroll.bind(this)}
+        renderItem={this.renderRow}
+        keyExtractor={(item, index) => String(index)}
         showsVerticalScrollIndicator={false}
         scrollEventThrottle={200}
-        onMoveShouldSetResponderCapture={() => {this.onListTouch(); return false;}}
-        keyExtractor={(item, index) => String(index)}
+        onMoveShouldSetResponderCapture={this.onMoveShouldSetResponderCapture}
+        onScroll={this.onScroll}
         refreshControl={this.props.refreshControl}
-        refreshing={this.props.refreshing || false}
+        refreshing={this.props.refreshing}
         onRefresh={this.props.onRefresh}
+        onScrollBeginDrag={this.props.onScrollBeginDrag}
+        onScrollEndDrag={this.props.onScrollEndDrag}
+        onMomentumScrollBegin={this.props.onMomentumScrollBegin}
+        onMomentumScrollEnd={this.props.onMomentumScrollEnd}
       />
     );
   }
 }
 
-export default ReactComp;
+export default ReservationList;
